+++ R/org_workflow_permissions.R
@@ -15,7 +15,10 @@ org_workflow_permissions = function(org) {
   res = purrr::safely(github_api_org_workflow_permissions)(org)
 
   if (failed(res)) {
-    cli::cli_abort("Failed to find org {.val {org}}.")
+    cli::cli_abort(c(
+      "Failed to retrieve default workflow permissions for org {.val {org}}.",
+      error_bullets(res)
+    ))
   }
 
   result(res)$default_workflow_permissions